requirements.txt
allure-pytest==2.15.2
pytest==9.0.2
selene==2.0.0rc9